move AbstractQueueAdmin#getConfigTableId to HBaseQueueAdmin#getConfigTableId
