+++ cdap-hbase-compat-0.98/src/main/java/co/cask/cdap/data2/transaction/queue/coprocessor/hbase98/HBaseQueueRegionObserver.java
@@ -16,7 +16,6 @@
 package co.cask.cdap.data2.transaction.queue.coprocessor.hbase98;
 
 import co.cask.cdap.common.queue.QueueName;
-import co.cask.cdap.data2.transaction.queue.AbstractQueueAdmin;
 import co.cask.cdap.data2.transaction.queue.ConsumerEntryState;
 import co.cask.cdap.data2.transaction.queue.QueueEntryRow;
 import co.cask.cdap.data2.transaction.queue.hbase.HBaseQueueAdmin;
@@ -89,7 +88,7 @@ public final class HBaseQueueRegionObserver extends BaseRegionObserver {
 
       conf = env.getConfiguration();
       String hbaseNamespacePrefix = HTable98NameConverter.getHbaseNamespacePrefix(tableDesc.getTableName());
-      TableId queueConfigTableId = AbstractQueueAdmin.getConfigTableId(namespaceId);
+      TableId queueConfigTableId = HBaseQueueAdmin.getConfigTableId(namespaceId);
       configTableNameBytes = HTable98NameConverter.toTableName(hbaseNamespacePrefix, queueConfigTableId).getName();
       configCache = ConsumerConfigCache.getInstance(conf, configTableNameBytes, new HTable98NameConverter());
     }
